net: add a sysctl to reflect the fwmark on replies
Kernel-originated IP packets that have no user socket associated
with them (e.g., ICMP errors and echo replies, TCP RSTs, etc.)
are emitted with a mark of zero. Add a sysctl to make them have
the same mark as the packet they are replying to.

This allows an administrator that wishes to do so to use
mark-based routing, firewalling, etc. for these replies by
marking the original packets inbound.

Tested using user-mode linux:
 - ICMP/ICMPv6 echo replies and errors.
 - TCP RST packets (IPv4 and IPv6).

[android-3.4 commit 3356997e1e1b2aa9959f046203e6d0b193bbd7f7]

Change-Id: If3dfa98be3115ef405366a90bb9514d453629baa
Signed-off-by: Lorenzo Colitti <lorenzo@google.com>
[steven@steven676.net: apply to android-3.0]